+++ src/com/amplitude/api/DeviceInfo.java
@@ -1,12 +1,5 @@
 package com.amplitude.api;
 
-import java.io.IOException;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Locale;
-import java.util.UUID;
-
 import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager.NameNotFoundException;
@@ -17,7 +10,14 @@ import android.location.LocationManager;
 import android.os.Build;
 import android.telephony.TelephonyManager;
 import android.text.TextUtils;
-import android.util.Log;
+
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+import java.util.UUID;
 
 public class DeviceInfo {
 
@@ -125,13 +125,15 @@ public class DeviceInfo {
             Location recent = getMostRecentLocation();
             if (recent != null) {
                 try {
-                    Geocoder geocoder = getGeocoder();
-                    List<Address> addresses = geocoder.getFromLocation(recent.getLatitude(),
-                            recent.getLongitude(), 1);
-                    if (addresses != null) {
-                        for (Address address : addresses) {
-                            if (address != null) {
-                                return address.getCountryCode();
+                    if (Geocoder.isPresent()) {
+                        Geocoder geocoder = getGeocoder();
+                        List<Address> addresses = geocoder.getFromLocation(recent.getLatitude(),
+                                recent.getLongitude(), 1);
+                        if (addresses != null) {
+                            for (Address address : addresses) {
+                                if (address != null) {
+                                    return address.getCountryCode();
+                                }
                             }
                         }
                     }
@@ -139,6 +141,8 @@ public class DeviceInfo {
                     // Failed to reverse geocode location
                 } catch (NullPointerException e) {
                     // Failed to reverse geocode location
+                } catch (NoSuchMethodError e) {
+                    // failed to fetch geocoder
                 }
             }
             return null;
@@ -185,9 +189,11 @@ public class DeviceInfo {
                 Method getId = advertisingInfo.getClass().getMethod("getId");
                 advertisingId = (String) getId.invoke(advertisingInfo);
             } catch (ClassNotFoundException e) {
-                Log.w(TAG, "Google Play Services SDK not found!");
+                AmplitudeLog.getLogger().w(TAG, "Google Play Services SDK not found!");
+            } catch (InvocationTargetException e) {
+                AmplitudeLog.getLogger().w(TAG, "Google Play Services not available");
             } catch (Exception e) {
-                Log.e(TAG, "Encountered an error connecting to Google Play Services", e);
+                AmplitudeLog.getLogger().e(TAG, "Encountered an error connecting to Google Play Services", e);
             }
             return advertisingId;
         }
@@ -279,7 +285,12 @@ public class DeviceInfo {
 
         List<Location> locations = new ArrayList<Location>();
         for (String provider : providers) {
-            Location location = locationManager.getLastKnownLocation(provider);
+            Location location = null;
+            try {
+                location = locationManager.getLastKnownLocation(provider);
+            } catch (IllegalArgumentException e) {
+                // failed to get last known location from provider
+            }
             if (location != null) {
                 locations.add(location);
             }